% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pscore.R
\name{bag_lapl_score}
\alias{bag_lapl_score}
\title{Bagging laplacian scoring for variables}
\usage{
bag_lapl_score(x, stand = FALSE, affin_method = "euclidean", bag_vol,
  max_core_num = getOption("max_core_num", 2L))
}
\arguments{
\item{x}{a data frame or matrix, row-wised input varible data for scoring}

\item{stand}{a logical value indicating if the varibles to be scaled}

\item{affin_method}{distance measure for affinity matrix calculation}

\item{bag_vol}{integer, sample volume in each bag iteration}

\item{max_core_num}{integer, number of cores/threads used for parallelized clustering}
}
\value{
a data frame with significance scores for the variables
}
\description{
Bagging laplacian scoring for variables
}

