% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pscore.R
\name{fisher_score}
\alias{fisher_score}
\title{Fisher scoring for variables}
\usage{
fisher_score(x, y, stand = FALSE, max_core_num = getOption("max_core_num",
  2L))
}
\arguments{
\item{x}{a data frame or matrix, row-wised input varible data for scoring}

\item{y}{an integer vector, response vector with one label for each row/compoonent of x}

\item{stand}{a logical value indicating if the varibles to be scaled}

\item{max_core_num}{integer, number of cores/threads used for parallelized clustering}
}
\value{
a vector with significance scores for the variables
}
\description{
Fisher scoring for variables
}

