% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pscore.R
\name{laplacian_score}
\alias{laplacian_score}
\title{Laplacian scoring for variables}
\usage{
laplacian_score(x, stand = FALSE, affin_method = "euclidean",
  max_core_num = getOption("max_core_num", 2L))
}
\arguments{
\item{x}{a data frame or matrix, row-wised input varible data for scoring}

\item{stand}{a logical value indicating if the varibles to be scaled}

\item{affin_method}{distance measure for affinity matrix calculation}

\item{max_core_num}{integer, number of cores/threads used for parallelized clustering}
}
\value{
a data frame with significance scores for the variables
}
\description{
Laplacian scoring for variables
}

